Calculate the power of two for large exponents
In many places the power of two was calculated from a too small value.

Integers that are of size `size_t` may be wider that `unsigned int`.
The expression `1u << order` is undefined for `order` >= `sizeof(unsigned int) * CHAR_BIT`.

Replace the instances of `1u << order` with an inline function that
instead shifts a `((size_t)1)`.

This allows `order < sizeof(size_t) * CHAR_BIT`.